Update README.md
The world of mortgage lending is ready for web3, DeFi and Dapps. I aim to make the first Dapp where the property owner has ownership(intended user rights) of their market value(appraised value) through completing their profile that is attached to their property(s), and controlling the information flow about said property(s). This can empower the owner to take a proactive approach to value and give them an annual value for tax and other purposes. Far too many people don't know the value of their real estate until it's time to do a loan on the property. This is the largest asset along with the largest loan most people will hold in their lifetime and should know and have control of the data that pertains to the value of the real estate. In the end, this will teach the owners the valuation principles as they interact with the Dapp. The Dapp will be able to upload/show pictures of improvements with scanned documents/invoices verifying the work has been completed by the owner. Other property improvements would be able to be added like a chain of title, but a chain of improvements(seller disclosures would possibly be a thing of the past).  This would greatly beneficial to all when obtaining a loan on the property, and much more efficient via blockchain technology, costing less money for each transaction and being super secure about personal data. This would incorporate the public sources of information, coupled with the current owner's input, minted and updated on the blockchain. Then this minted information could be passed to a new owner at time of sale. Just think... If a property has been sold 3 times in 5 years the amount of information those 3 different owners can contribute. Property inspection reports, improvements by the owners, 3 different seller disclosures to refer back to. No more questions of the history of the property. Each new owner gets the background and also gets to put their mark on the lifespan of that property. If it's owned by Bruce Willis, there's no question or doubt that it was, and can be verified. No more shady real estate deals, valuations, or agents. Real estate professionals should be driven by one motive and one motive only, to promote public trust. Ever since the ability to buy and sell real estate to the masses there have been shady people looking to take advantage of/be predatory to uneducated buyers and sellers. This Dapp can solve many of these shady sales and work to promote public trust to the real estate properties, transactions and the people in the industry that truly want people to be educated homeowners to bridge the wealth divide. By bringing all things about the property to light, the questions of what is valuable and not about a property can be answered for everyone.